Include positions in AST, remove SourceMap
Although I liked that the AST was not cluttered with metadata in the
SourceMap model, it had a problem where two structurally equal nodes of
the AST could not be distinguished in the SourceMap. A solution is to add
some extra information in these nodes, so we might then just as well add
the location in there.